Manually change type of nDiff to float, for normal precision on devide this values
Now it move cursor really granular

+++ src/qwebpage.cpp
@@ -100,9 +100,9 @@ namespace Wrapper {
 			}
 
 			if (xDiff < yDiff) {
-				ratioY = xDiff ? (yDiff / xDiff) : 0;
+				ratioY = xDiff ? ((float)yDiff / (float)xDiff) : 0;
 			} else {
-				ratioX = yDiff ? (xDiff / yDiff) : 0;
+				ratioX = yDiff ? ((float)xDiff / (float)yDiff) : 0;
 			}
 
 			maxLoopIterations = ratioX ? (xDiff / ratioX) : (yDiff / ratioY);
